[DNC] feat(core) pluggable delayed_response callback
This commit alters the core access logic introduced in dcc0ce8
to provide a pluggable callback function for plugins that synthesize a
response in the access phase. We default to the predefined
'responses.flush_delayed_responses' to maintain existing behavior,
so this introduces no behavior change in existing plugins. We also
demonstrate an explicit definition of the callback in a plugin that does
not use 'kong.tools.responses' (the request-termination plugin).

This commit also changes the priority of the UDP logging plugin to allow
for udp-log to demonstrate this issue, though this is out of scope and
only tangentially relevant (I was testing locally with UDP logging).